toolkit/components/url-classifier/tests/test_bug356355.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=356355

This is a port of all the existing EnchashDecryptor unittests to the
mochitest framework.
-->
<head>
  <title>Test for Bug 356355</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=356355">Mozilla Bug 356355</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

// These have been moved to test_enchash_decrypter.xhtml
// We have this placeholder because mochitest tinderbox don't clobber
// so removed files still try to run.
ok(true, 'tests moved to test_enchash_decrypter.xhtml');

]]>
</script>
</pre>
</body>
</html>
